Handling Oracle's missing ILIKE operator
Oracle does not have the ILIKE operator, which makes the pure usage of `wlNext.caseSensitive` flag not possible.

Here's what I found out and what my change is based on. Setting `wlNext.caseSensitive: true` produces a plain LIKE query, which makes perfect sense.

Setting `wlNext.caseSensitive: false` produces an ILIKE query with the LOWER wrap, which is a bit confusing to me, because the ILIKE is case-insensitive, and therefore there should be no need for LOWER. Never the less, this case throws an error when using with Oracle, because Oracle does not have the ILIKE operator.

My solution above was twofold. First, I fixed the DRY violation on the `wlNext.caseSensitive` flag checking. Second, I added a new flag `wlNext.forceLike` with which one can indicate that the database requires LIKE to be used, instead of ILIKE.